pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testops-insights"
version = "0.1.0"
description = "TestOps Dashboard - Analyze test results to identify flaky tests and assess pipeline health"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
testops-insights = "testops_insight.cli.main:main"

